MRCA/requirements.txt
numpy>=1.19.2
pandas>=1.2.4
torch>=1.8.0
scikit-learn>=0.24.2
networkx>=2.5
matplotlib>=3.4.2
pyyaml>=5.4.1
drain3>=0.9.1  # 用于日志解析